Reintroduce separate date picker
Separate date and time pickers were removed in commit 0666a1b in favor of
a custom 'format' field in the datetime picker. However, these formats may
be a bit of a mystery for the less tech-savvy and so I've reimplemented a
separate date picker to provide a more user-friendly way of letting users
select either picker depending on their needs.

This also means that the 'format' field has been deprecated. Consequently,
datetime combinations are stored as a Unix timestamp (in string format).