Introduce new 'autocomplete' and 'skip' features
Auto-completion is a way to handle ambiguous or non-deterministic input.
A good example of this are questions that require a volunteer to enter
the name of a location, a location that could be written in more than one
way. By enabling auto completion, the input field is filled with the
expected answer.

Like auto-completion, the skipping mechanism allows us to skip questions
that may not be pertinent during the tutorial. However, skipping should
be preferred over auto-completion in cases where one can not reliably
determine an expected answer, most notably geotagging-type questions.

Additionally, the geotagx.questionnaire.questionChanged_ handler has been
added and is called each time a different question is presented to the
user. In our case, it is used to auto-complete a question as soon as
possible.