Remove reference to a second photo in the tutorial
Currently, most (if not all) project tutorials only have one image to
analyse and as such, the "completed one of the project's tutorials (...)"
message that is displayed upon completion of a tutorial wrongly implies
the existence of more than one image. The message has been changed to
alleviate confusion.

References geotagx/geotagx-theme#95.