Make minor UX improvements
The 'Borders' layer adds region borders and text to satellite and aerial
images with the aim of making it easier to navigate the aforementioned
imagery. Past a certain resolution however, the overlayed text becomes
more of an obstruction.

The layer is now hidden when we go past this resolution, and made visible
when it is not considered to be a hindrance.