Improve OpenLayers search functionality
The search functionality previously centered the map at a location's
longitude and latitude. The zoom factor never changed which became
problematic when users searched for very specific locations, e.g.
streets or building addresses, because while the map was correctly
centered at the location, it did not automatically zoom in to provide
more detail.

The search functionality now centers the map around the location's
bounding box, i.e. a rectangle within which the location is encompassed,
effectively using a variable zoom factor based on the bounding box's area.

References #15.