Add 'subject-type' field to project configuration
This parameter is meant to supersede the 'PDF mode' that allows a project
creator to set the type of data the project will be analyzing.
The current implementation of the 'PDF mode' doesn't allow for multiple
projects to be built correctly in batches since once the 'pdfmode' flag
is set to true, all projects are assumed to have PDFs as their subject
type. Adding the parameter to the project configuration solves this
little pickle.

Furthermore, this solution should also make it easier to introduce new
subject types (e.g. audio, video) in the future.